Bring update code back into synch across the various versions. The confirmation screen is only showing what needs to be done, and if no updates are needed a message is displayed.
Due to the interplay between the original forms, the start array that the submission handler recieves does not have any entries for modules that do not need to be updated. The code now correctly populates the $start array, which was previously executing unneeded functions.

I discovered that the install profile global is not properly populated in d6 and d7, which causes modules stored there to be disabled during the upgrade process. I now initialize them properly, and have added additional logging to those parts of the code